requirements.txt
# Core dependencies
anthropic>=0.39.0        # Claude API
python-dotenv>=1.0       # Environment variables

# Speech-to-Text
openai-whisper          # Local STT
SpeechRecognition>=3.10 # Alternative STT with Google support

# Text-to-Speech
gtts>=2.5               # Google Text-to-Speech
pyttsx3>=2.90           # Offline TTS (backup)

# Audio processing
pyaudio>=0.2.14         # Audio I/O
numpy>=1.24             # Audio processing
sounddevice>=0.4.6      # Alternative audio I/O

# Wake word detection
pvporcupine>=3.0        # Wake word detection engine
pvrecorder>=1.2         # Audio recording for Porcupine

# Utilities
colorama>=0.4.6         # Colored terminal output
requests>=2.31          # HTTP requests